feat: add user-delete feature

+++ pkg/api-server/api_v1/handlers/user.go
@@ -31,10 +31,11 @@ func UserSignUp(c *gin.Context) {
 		return
 	}
 
-	sqlStr := "INSERT INTO suglider.user_info(UNHEX(REPLACE(UUID(), '-', '')), username, password, mail, address) VALUES (?,?,?,?)"
+	sqlStr := "INSERT INTO suglider.user_info(user_id, username, password, mail, address) VALUES (UNHEX(REPLACE(UUID(), '-', '')),?,?,?,?)"
 	_, err = mariadb.DataBase.Exec(sqlStr, request.Username, request.Password, request.Mail, request.Address)
 	if err != nil {
 		log.Println("Insert user_info table failed:", err)
+		c.JSON(http.StatusBadRequest, gin.H{"message": "User create failed"})
 	} else {
 		c.JSON(http.StatusOK, gin.H{"message": "User created successfully"})
 	}
@@ -51,19 +52,22 @@ func UserDelete(c *gin.Context) {
 	}
 
 	if request.User_id == "" {
-		sqlStr := "DELETE FROM suglider.user_info WHERE username=?, mail=?"
+		sqlStr := "DELETE FROM suglider.user_info WHERE username=? AND mail=?"
 		_, err := mariadb.DataBase.Exec(sqlStr, request.Username, request.Mail)
 		if err != nil {
 			log.Println("Delete user_info data failed:", err)
+			c.JSON(http.StatusBadRequest, gin.H{"message": "User delete failed"})
 		} else {
 			c.JSON(http.StatusOK, gin.H{"message": "User deleted successfully"})
 		}
 	
 	} else {
-		sqlStr := "DELETE FROM suglider.user_info WHERE user_id=?, username=?, mail=?"
+		// UNHEX(?) can convert user_id to binary(16)
+		sqlStr := "DELETE FROM suglider.user_info WHERE user_id=UNHEX(?) AND username=? AND mail=?"
 		_, err := mariadb.DataBase.Exec(sqlStr, request.User_id, request.Username, request.Mail)
 		if err != nil {
 			log.Println("Delete user_info data failed:", err)
+			c.JSON(http.StatusBadRequest, gin.H{"message": "User delete failed"})
 		} else {
 			c.JSON(http.StatusOK, gin.H{"message": "User deleted successfully"})
 		}
